fix: lint

+++ modules/frontend/auth/data/src/services/interceptors.ts
@@ -12,8 +12,7 @@ export function configAxiosInstance(instance: AxiosInstance) {
     (error) => {
       console.warn({ error });
 
-      if (error.response?.status === 401)
-        authCookiesService.deleteCookieClient();
+      if (error.response?.status === 401) authCookiesService.deleteCookieClient();
 
       return Promise.reject(error);
     }
@@ -21,12 +20,15 @@ export function configAxiosInstance(instance: AxiosInstance) {
   return instance;
 }
 
-export function handleAxiosError(functionName: string, error: Error) {
-  console.warn(`[API] AuthService.${functionName}`, error);
-}
-
-export function handleUnexpectedError(functionName: string, error: Error) {
-  const message = `[JS] AuthService.${functionName}`;
-  console.error(message, error);
-  return Error(message);
+export function Log(context: string) {
+  return {
+    apiError(functionName: string, error: Error) {
+      console.warn(`[API] ${context}.${functionName} : ${error.name} - ${error.message} \n\n${error.stack}\n\n`, error);
+    },
+    unexpectedError(functionName: string, error: Error) {
+      const message = `[JS] AuthService.${functionName} : ${error.name} - ${error.message} \n\n${error.stack}`;
+      console.error(message, error);
+      return Error(message);
+    },
+  };
 }
